requirements.txt
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
dnspython==2.2.1
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.9
user-agent==0.1.10
yarl==1.7.2
